Add support for terminal default colors
At startup, ncurses initializes the standard color pair 0 with black
and white as the default colors, independent of the colors used by the
terminal.

If use-default-colors is t, the screen is initialized with the colors
defined by the terminal, including transparent backgrounds.

The foreground and background colors pre-defined by the terminal are
both named :default.

If use-default-colors is nil, the fg :default is white, and bg black.