Modify how t08c displays complex chars
Do not use write-char and format ~C for complex-chars.
Even if they work with complex-chars, they may not in the future.
Only lisp characters should be used.

Complex chars can be used with princ and format ~A, which use
print-object underneath, which can safely be customized to display
rendered complex chars.