Add support for function/callback menu items and menu names
Add a slot name for menus.

Add a boolean option to the title of decorated-windows and menu-windows,
to be able to turn off the display of titles.

If title is t, the name of the menu is displayed as its title.
If the title is a string, it overrides the menu name.