Add window slots to fields and forms, refactor draw functions
Remove the window parameter from draw functions.

The draw functions now assume that the drawn objects either has a window
slot, or is part of a form which has a window slot.

If part of a form, every element (field) will get a pointer to the parent
form upon form initialization.

The clear function is now a method, specialized on windows and fields.