Add utility functions complete-pair and funcall-make-chtype
complete-pair allows window-related colors to be passed as single
colors, either foreground or background.

The missing color is completed from a window's color-pair property,
background or from the ncurses default color pair.

funcall-make-chtype does the same as funcall-make-cchar_t, it takes
components of a complex char, assembles the underlying ncurses
chtype (or cchar_t), and then calls the passed function on it.