Refactor the initialisation of the default color pair 0
The ncurses default pair 0 is white on black.

When the window color is set to some other pair, the default color
pair is ignored by several functions, including addch.

Therefore (:white :black) can't be associated with the pair 0.

From now, the default pair 0 is (:default-fg :default-bg).

What color are really used can be read and set by the screen
accessor default-color-pair.

use-default-colors-p is renamed to the more proper use-terminal-colors-p

The fg and bg color set by the terminal can be accessed through
the special color :terminal (color number -1).

This color is only available after it is initialized by
use-terminal-colors-p set to t or default-color-pair set to
(:terminal :terminal).

Due to a bug with ncurses, the default color pair con not contain
only one terminal color. As soon as one is terminal, the other is also
automatically set to terminal.

The special color :terminal allows transparency to be used with
ncurses, by setting, for example a transparent background in the
terminal, and then setting a window background to :terminal.