Add WITH-TERM macro for cross-thread evaluation
WITH-TERM uses a thread-safe FIFO queue to queue up requests which
should be evaluated inside the terminal thread. For this to work one
has to call IDLE from inside the terminal thread to pop requests from
the FIFO and evaluate them.
When a condition is signaled while the body of WITH-TERM is evaluated,
it is handled by IDLE and put inside a WITH-TERM-ERROR which also
contains the failed form. The condition is then signaled again from
within a restart which allows skipping the failed form and continue
evaluating requests.