Process *JOB-QUEUE* in event-loop
Instead of making the user call CROATOAN:PROCESS, call it from the
event-loop. This only works when input-blocking is not T, hence either
set to a timeout or NIL. This will not override the NIL event, which
will still be dispatched so that the user can handle it.
This should not change behaviour besides making CROATOAN:SUBMIT work
once the ncurses terminal thread has entered the event-loop and
input-blocking is not set to T.