Replace the define-keymap function with a convenience macro
The goal is to replace the syntax to look more like case or cond:

(define-keymap checkbox-map
  (#\space 'toggle-checkbox)
  (#\x 'toggle-checkbox))

instead of being passed as a plist.

(define-keymap 'checkbox-map
  (list
    #\space 'toggle-checkbox
    #\x 'toggle-checkbox))

The new macro allows control chars to be given in caret notation,
for example "^A" instead of #\soh.

The plist can still be passed as an argument when a keymap object is
initialized.