Bugfix: Do not complete a nil color pair from the background char
This restores the behavior expected from ncurses:

If a simple space character (without color) is added to a window with
a set background char, ncurses displays the background char.

If a colored space is added to a window with a set background char,
ncurses displays the colored space.

This allows the clear function to overwrite a window with the space char,
and have the background char used instead of space.