Remove the frame-rate widget slot
frame-rate is now an additional accessor for the input-blocking window
property.

The ncurses blocking delay is set which allows an immediate response
if an event is available.

sleep is not called during the event loop any more, since it causes an
input lag.

Allowed values: nil or a positive integer denoting how often the nil
event will be returned per second when no other event is available.