Add a panel class to replace decorated-window
* Rename panel (decorated-window) to extended-window

An extended-window is a window with a sub-window for content,
for example menu items of form elements.

* Add a new panel class

A panel is a window with a dedicated border window and a
background window to represent a shadow.

The purpose of a panel is that it can be used like a normal
window (add, move, clear) and that the border and shadow can
be styled independently of the content window.

For an extended window (menu-window, form-window) the style
of the content window is also the style of the border and
using the normal window functions (add, move, clear) is
awkward because we have to specifically target the "inner"
sub-window.

* Add style support for windows

:foreground
  :fgcolor
  :bgcolor
  :attributes
:background
  :fgcolor
  :bgcolor
  :attributes
  :simple-char

In a panel window, the :border and :shadow windows can be styled
separately.

* Base the menu-window on extended-window again, remove menu-panel

There is no real use for a menu displyed in a simple window.

The simplest use case is a menu displayed in an extended window
with or without a border.

The menu-window is used as a submenu and thus will explicitely
not be able to display a title because submenu titles dont look good.

If you want a menu with a title, the menu-panel based on the new
panel class should support more elaborate decorations like titles,
a scroll bar, etc. menu-panel is not available yet.