- fixed array-slice;
The function was allocating a results  vector as big as the input when
the parameter 'end' was null.

This would gives wrong results when the parameter 'start' was > 0 (the
majority of the call to this function is called, i think).

This patch try to takes into  account the parameter 'end' when is null
and allocates a correctly sized resulting array.