Add a callback-type slot to components
Currently supported callback types are :object and :object-event (default).

This allows the event handling routines to accept handlers in the form

(lambda (object) ...)

in addition to the initially supported

(lambda (object event) ...)

Many handlers do not need the event to be passed as an argument.

Test setting the callback type in matrix4.