Update mouse support to work with bind/run-event-loop
To make this feasible, a backward-incompatible change of the event
handling parts had to be made.

get-event and get-wide-event now return an event object instead of a
simple character or keyword, so that mouse click positions can be passed
as slots.

Mouse events can now be bound to handlers like any other event:

(bind win :button-1-click (lambda (w e) ...))

The mouse position variables in (event-case (win event y x)) are not
needed any more and the form can be reduced to (event-case (win event)).

The underlying ncurses mouse events have been updated to
NCURSES_MOUSE_VERSION 2, which requires ncurses 6 to be built with
--enable-ext-mouse. This enables the use of scrolling with mouse
wheels and touch pads, by binding scrolling movement to buttons 4 and 5.