Allow modifier keys to be returned with mouse events
Add a slot modifiers and accessor event-modifiers to mouse-events.

If any of the modifiers CTRL, SHIFT and ALT are pressed during a mouse
event, they will be returned as a list in the slot modifiers, and
can be checked in user code.

Allow :all-mouse-events to be passed to set-mouse-event, to activate
all mouse events except :report-mouse-position, tested in t14c1.