Allow a window layout to be automatically calculated
If windows are given as children of a row or column layout,
their position and dimensions do not have to be hard-coded,
but can be calculated to fill the screen.

Windows are passed as leaves of a layout tree in form of a plist,
and the objects are initialized after the geometry has been
calculated.

Added a generic function (accessor) geometry.

Added functions calculate-layout, initialize-leaves.

The function leaves returns a list of windows in a layout tree.

Added example t43, which shows a succint way to specify a window
layout for a mock chat application.