Fix textarea behavior when a newline is added to a wrapped line
After a full line of characters, the cursor is wrapped to the next line.

Adding a newline character at that point now does not move the cursor
down to the third line, but stays in the wrapped, second line.

This behavior corresponds to that of Vim.

Emacs does not even allow a character to be displayed in the last column,
but adds the contination char '\' instead and moves the typed char to the
first column, so that the cursor actually can not appear in the first
column at all.

The Vim behavior is more intuitive and more appropriate for a simple
text widget.

A second minor bug was fixed where the textarea was not able to accept
and display wide characters.