Add border, padding and table for menus within forms
Allow a border and padding to be set for menus that are used as form
elements. The properties are not yet supported for menu-windows.

Add :table t to draw table lines between menu items.

The table is drawn using ACS box drawing chars, other drawing styles
will be added later.

The properties border-width and padding are only in effect for
borders, not for tables.

Base menu class on layout class to get support for row and column gaps.

Remove function draw-menu-item, refactor it into draw-item.

Add column-widths and row-heights methods for arrays.

Add ref2d, function to get i,j from a list like aref.