Add a cyclic property to collection lists
The default behavior is not cyclic, but the forms default is cyclic.

Check the cyclic property for collections and layouts.

Add helper functions previous-item-p, next-item-p, last-item-number.

Bugfix:

Calling clear on a menu-panel tries to clear the drawn menu, but we
actually want to clear the window.

We can fix that by reversing the parent class precedence from (menu
panel) to (panel menu).

Bugfix:

When scrolling is enabled for a menu in a menu-panel, it at the same
time enables scrolling of the panel window itself, and we dont want
that.

We fix this by checking for a menu when calling ncurses:scrollok
during window initialization.