Add a way to align left/right and ellipsize menu item titles
All the initarg :align, so that items can be aligned if they are
shorter than the max-item-length or the calculated column width.

Do not allow the length of the item title to exceed max-item-length,
in that case the title is ellipsized or truncated.

Add the initarg :ellipsis, where the string can be specified,
which will replace the excess characters of a too long title.
If the ellipsis is an empty string, the title will simply be
truncated.

This patch resolves the issue #39 opened on 210213 by Mr-X-666.